+++ src/bytecode.rs
@@ -163,8 +163,9 @@ pub enum Op {
     DefineClassAccessor(u8), // same but enumerable=false (for class methods)
     GetProp,
     SetProp,
-    DefineMethod, // define non-enumerable method property: stack [obj, key, value]
-    GetElem,      // computed member
+    DefineDataProperty, // define enumerable own data property: stack [obj, key, value]
+    DefineMethod,       // define non-enumerable method property: stack [obj, key, value]
+    GetElem,            // computed member
     SetElem,
     DeleteProp,
     SetProto,        // pop [proto, obj]; set obj's [[Prototype]] to proto